pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "athena-regime"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "matplotlib>=3.8",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "pyarrow>=14.0",
    "duckdb>=1.1",
    "pyyaml>=6.0",
    "requests>=2.31",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov"]
calendars = ["pandas-market-calendars>=4.3"]
bloomberg = ["openpyxl>=3.1"]

[project.scripts]
athena = "athena_regime.cli:main"
data = "data.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
src = ["src"]
